+++ tests/e2e/tests/cypress/e2e/migration/import.cy.js
@@ -131,7 +131,7 @@ describe('Migration System - Import', () => {
 
         cy.exec(`cp ${fixturesDir}/MigrationFiles/${file} ${baseDir}/${file}`);
         cy.exec('docker exec application php index.php db:migrate --dry -f');
-        cy.visit("/migration/checkImport");
+        cy.visit({ url: "/migration/checkImport", failOnStatusCode: false });
 
         cy.contains("Table 'app.migration_import' doesn't exist");
 
@@ -143,7 +143,7 @@ describe('Migration System - Import', () => {
 
         cy.exec(`cp ${fixturesDir}/MigrationFiles/${file} ${baseDir}/${file}`);
         cy.exec('docker exec application php index.php db:migrate -f');
-        cy.visit("/migration/checkSkippedMigrations");
+        cy.visit({ url: "/migration/checkSkippedMigrations", failOnStatusCode: false });
 
         cy.contains("Table 'app.migration_skip' doesn't exist");
 
@@ -182,7 +182,7 @@ describe('Migration System - Import', () => {
 
         cy.exec(`cp ${fixturesDir}/MigrationFiles/${file} ${baseDir}/${file}`);
         cy.exec('docker exec application php index.php db:migrate -e production -f');
-        cy.visit("/migration/checkEnvMigrations");
+        cy.visit({ url: "/migration/checkEnvMigrations", failOnStatusCode: false });
 
         cy.contains("Table 'app.migration_env' doesn't exist");
 
@@ -205,7 +205,7 @@ describe('Migration System - Import', () => {
 
         cy.exec(`cp ${fixturesDir}/MigrationFiles/${file} ${zubzetMigrationPath}/${file}`);
         cy.exec('docker exec application php index.php db:migrate --exclude-external -f');
-        cy.visit("/migration/checkImport");
+        cy.visit({ url: "/migration/checkImport", failOnStatusCode: false });
 
         cy.contains("Table 'app.migration_import' doesn't exist");